+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## [Unreleased]
 
+## [0.2.1] - 2021-07-24
+- Add `CallableTree::Node::root?`.
+- Add `CallableTree::Node::Internal#seek!` that make destructive change.
+- Add `CallableTree::Node::Internal#broadcast!` that make destructive change.
+- Add `CallableTree::Node::Internal#compose!` that make destructive change.
+
 ## [0.2.0] - 2021-06-15
 - Change `CallableTree::Node::Internal#append` to return a new instance.
   To keep the same behavior as the older version, use `CallableTree::Node::External#append!` that make destructive change.